EHR-12345 Flag Summary story books
